+++ CHANGELOG.md
@@ -11,4 +11,5 @@
 9. Installed console.table package
 10. Added some seed data to HRIS Database
 11. Tested List All Employees by Salary showing Role, Salary and Department
-12. Created View All Employees query, called from server.js and formatted using console.table 
+12. Created View All Employees query, called from server.js and formatted using console.table
+13. Added queries for View All Departments and View All Roles